requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.2
billiard==4.1.0
celery==5.3.1
click==8.1.4
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==4.0.10
django-cors-headers==4.2.0
django-filter==23.2
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.3.0
kombu==5.3.1
Markdown==3.4.3
mongoengine==0.27.0
prompt-toolkit==3.0.39
pymongo==3.12.1
python-dateutil==2.8.2
pytz==2023.3
redis==4.6.0
six==1.16.0
sqlparse==0.2.4
typing_extensions==4.7.1
tzdata==2023.3
vine==5.0.0
wcwidth==0.2.6
